+++ CHANGELOG.txt
@@ -1,5 +1,9 @@
 v.5.4.0
 - Added Threat Rolls, alternate Load Levels, and Experience clocks in support of Deep Cuts supplements (#19)
+- Deduplicated listeners and related function by moving them to blades-sheet.js
+- Opening a contact that is not imported to the world will now import the actor and connect the imported actor to the character sheet
+- Radio boxes now toggle when clicked and the zero buttons have been removed
+- All multiboxes on scoundrel and crew sheets are dynamically generated and can be modified through active effects
 
 v.5.3.2
 - Add all rolls type upon attribute click (#23 @nic-ttrpg)
