+++ styles/blades.css
@@ -1464,7 +1464,7 @@
 .blades-in-the-dark.actor.pc * .teeth-experience label {
   height: 30px;
   width: 10px;
-  background-image: url("assets/teeth/stresstooth-red.png");
+  background-image: url("assets/teeth/xptooth-red.png");
   background-repeat: no-repeat;
   background-size: contain;
   margin-right: 5px;
@@ -1494,7 +1494,7 @@
   background-image: url("assets/teeth/stresstooth-halfgrey.png");
 }
 .blades-in-the-dark.actor.pc * .teeth-experience input:checked + label {
-  background-image: url("assets/teeth/stresstooth-red.png");
+  background-image: url("assets/teeth/xptooth-red.png");
 }
 .blades-in-the-dark.actor.pc * .harm-armor {
   display: flex;
@@ -1564,6 +1564,7 @@
   display: flex;
   flex-direction: row;
   justify-content: space-between;
+  /* Hide the browser's default checkbox */
 }
 .blades-in-the-dark.actor.pc * .attributes {
   display: flex;
