+++ system.json
@@ -2,7 +2,7 @@
   "id": "blades-in-the-dark",
   "title": "Blades in the Dark",
   "description": "Blades in the Dark game system.",
-  "version": "5.3.2.3",
+  "version": "5.3.2.4",
   "compatibility": {
     "minimum": 11,
     "verified": "12.330"
