+++ templates/actor-sheet.html
@@ -8,14 +8,14 @@
 
     <div class="grow-two">
       <div id="character-{{_id}}-name-box">
-        <label for="character-{{_id}}-name">{{localize "BITD.Name"}}</label>
+        <label for="character-{{_id}}-name"><b>{{localize "BITD.Name"}}</b></label>
         <input type="text" id="character-{{_id}}-name" class="name" name="name" value="{{name}}">
       </div>
 
       <div id="character-{{_id}}-traits" class="flex-column">
 
         <div class="item-block flex-horizontal">
-          <div class="item-class-label"><a class="item-add-popup" data-distinct="true" data-item-type="heritage">{{localize "BITD.Heritage"}}</a></div>
+          <div class="item-class-label"><a class="item-add-popup" data-distinct="true" data-item-type="heritage"><b>{{localize "BITD.Heritage"}}</b></a></div>
           {{#each items as |item id|}}
               {{#if (eq item.type "heritage")}}
               <div class="item flex-horizontal" data-item-id="{{item._id}}">
@@ -31,7 +31,7 @@
         </div>
 
         <div class="item-block flex-horizontal">
-          <div class="item-class-label"><a class="item-add-popup" data-distinct="true" data-item-type="background">{{localize "BITD.Background"}}</a></div>
+          <div class="item-class-label"><a class="item-add-popup" data-distinct="true" data-item-type="background"><b>{{localize "BITD.Background"}}</b></a></div>
           {{#each items as |item id|}}
             {{#if (eq item.type "background")}}
             <div class="item flex-horizontal" data-item-id="{{item._id}}">
@@ -47,7 +47,7 @@
         </div>
 
         <div class="item-block flex-horizontal">
-          <div class="item-class-label"><a class="item-add-popup" data-distinct="true" data-item-type="vice">{{localize "BITD.Vice"}}</a></div>
+          <div class="item-class-label"><a class="item-add-popup" data-distinct="true" data-item-type="vice"><b>{{localize "BITD.Vice"}}</b></a></div>
           {{#each items as |item id|}}
               {{#if (eq item.type "vice")}}
               <div class="item flex-horizontal" data-item-id="{{item._id}}">
@@ -63,7 +63,7 @@
         </div>
 
         <div class="item-block flex-horizontal">
-          <div class="item-class-label"><a class="item-add-popup" data-distinct="true" data-item-type="class">{{localize "BITD.Class"}}</a></div>
+          <div class="item-class-label"><a class="item-add-popup" data-distinct="true" data-item-type="class"><b>{{localize "BITD.Class"}}</b></a></div>
             {{#each items as |item id|}}
                 {{#if (eq item.type "class")}}
                 <div class="item flex-horizontal" data-item-id="{{item._id}}">
@@ -79,7 +79,7 @@
           </div>
 		
 		<div class="item-block flex-horizontal">
-          <div class="item-class-label"><a class="item-add-popup" data-distinct="true" data-item-type="crew">{{localize "BITD.CrewName"}}</a></div> 
+          <div class="item-class-label"><a class="item-add-popup" data-distinct="true" data-item-type="crew"><b>{{localize "BITD.CrewName"}}</b></a></div> 
           {{#each system.crew as |item id|}}
             {{#unless (eq item.name undefined)}}
               <div class="item flex-horizontal" data-item-id="{{item.id}}" data-item-type="{{item.type}}">
@@ -102,7 +102,7 @@
 
     <div class="grow-two">
       <div id="character-{{_id}}-alias-box">
-        <label for="character-{{_id}}-alias">{{localize "BITD.Alias"}}</label>
+        <label for="character-{{_id}}-alias"><b>{{localize "BITD.Alias"}}</b></label>
         <input type="text" id="character-{{_id}}-alias" class="alias" name="system.alias" value="{{system.alias}}">
       </div>
 
@@ -145,12 +145,13 @@
       {{#multiboxes system.stress.value}}
 
       <input type="radio" id="character-{{_id}}-stress-0" name="system.stress.value" value="0" dtype="Radio">
-      <label class="black-label" for="character-{{_id}}-stress-0">{{localize system.stress.name}}</label>
+      <!--<label class="black-label" for="character-{{_id}}-stress-0">{{localize system.stress.name}}</label>-->
+	  <div class="black-label">{{localize system.stress.name}}</div>
 
       {{#times_from 1 system.stress.max}}
         <input type="radio" id="character-{{../_id}}-stress-{{this}}" name="system.stress.value" value="{{this}}"
                dtype="Radio">
-        <label for="character-{{../_id}}-stress-{{this}}"></label>
+        <label class="radio-toggle" for="character-{{../_id}}-stress-{{this}}"></label>
       {{/times_from}}
 
       {{/multiboxes}}
@@ -159,7 +160,8 @@
     <div id="character-{{_id}}-trauma-container" class="small-teeth-container">
       {{#traumacounter system.trauma.list}}
       <div id="character-{{_id}}-trauma" class="small-teeth-wrap">
-        <label class="black-label" for="character-{{_id}}-trauma-counter-0">{{localize system.trauma.name}}</label>
+        <!--<label class="black-label" for="character-{{_id}}-trauma-counter-0">{{localize system.trauma.name}}</label>-->
+		<div class="black-label">{{localize system.trauma.name}}</div>
         <div id="character-{{_id}}-trauma-teeth" class="small-teeth">
           <input type="radio" id="character-{{_id}}-trauma-counter-0" name="system.trauma.value" value="0">
 
@@ -294,16 +296,19 @@
         <div class="label-stripe">
           <p><label for="character-{{_id}}-experience-0">{{localize "BITD.Abilities"}}</label></p>
           <p><a class="item-add-popup" data-item-type="ability" data-tooltip="{{localize 'BITD.AddAbility'}}"><i class="fas fa-plus-square"></i></a></p>
-          <div id="character-{{_id}}-experience" class="teeth-experience">
-            <label for="character-{{_id}}-experience-0"><i class="fab fa-creative-commons-zero nullifier"></i></label>
+          {{#if (getSetting 'ClockXP')}}
+		  {{else}}
+		  <div id="character-{{_id}}-experience" class="teeth-experience">
+            <!--<label for="character-{{_id}}-experience-0"><i class="fab fa-creative-commons-zero nullifier"></i></label>-->
             {{#multiboxes system.experience}}
             <input type="radio" id="character-{{_id}}-experience-0" name="system.experience" value="0" dtype="Radio">
             {{#times_from 1 system.experience_max}}
             <input type="radio" id="character-{{_id}}-experience-{{this}}" name="system.experience" value="{{this}}" dtype="Radio">
-            <label for="character-{{_id}}-experience-{{this}}"></label>
+            <label class="radio-toggle" for="character-{{_id}}-experience-{{this}}"></label>
             {{/times_from}}
             {{/multiboxes}}
           </div>
+		  {{/if}}
         </div>
 
         <div id="character-{{_id}}-abilities-list">
